+++ src/components/FormInputs/FormInputs.jsx
@@ -16,7 +16,8 @@
 
 */
 import React, { Component } from "react";
-import { FormGroup, ControlLabel, FormControl, Row } from "react-bootstrap";
+import {Form, FormGroup, ControlLabel, FormControl, Row, } from "react-bootstrap";
+
 
 function FieldGroup({ label, ...props }) {
   return (