+++ src/index.js
@@ -38,7 +38,7 @@ ReactDOM.render(
     <BrowserRouter>
       <Switch>
         <Route path="/admin" render={props => <AdminLayout {...props} />}/>
-        <Redirect from="/" to="/admin/dashboard"/>
+        <Redirect from="/" to="/admin/noticias"/>
       </Switch>
     </BrowserRouter>
   </Provider>,